Add tablet version and more features
Added tablet version of site. Changed nav bar code so that the buttons are generated from a list instead of hard-coded. Changed class names for clarity purposes. Removed blur effect on the description on the crew page. Added transitions to the nav buttons in the header and on the destination page. Added Sass mixin to style the body of text on each page for the tablet version.