Add desktop version and make small changes
Added styling for the desktop version of the site which kicks in at 1440px.

Changed the way that the body text is styled for each page of the website.
Previously, it was styled using Sass mixins. Now they're styled using CSS
custom properties.

Removed blurring effect added using the backdrop-filter property on the
body text of the crew page as it was unnecessary.

Added a <p> tag to the menu button on the mobile version of the website
for screenreaders.

Edited spacesport image on the technology page to crop out the unwanted
blue background